feat: separated app into insights and schedule
