+++ src/App.tsx
@@ -1,10 +1,4 @@
-import { useState, useEffect, useCallback } from "react";
-import {
-  DragDropContext,
-  Droppable,
-  Draggable,
-  DropResult,
-} from "@hello-pangea/dnd";
+import { useState, useEffect } from "react";
 import { Button } from "./components/ui/button";
 import { Input } from "./components/ui/input";
 import {
@@ -12,91 +6,50 @@ import {
   EnvelopeClosedIcon,
   CalendarIcon,
 } from "@radix-ui/react-icons";
-import { StatsPanel } from "./components/StatsPanel";
 import { mockMeetings } from "./mockData";
-import type { Meeting, MeetingStats } from "./types";
-import { MeetingCard } from "./components/MeetingCard";
 import { sendEmail } from "@/services/emailService";
 import { useSettingsStore } from "./stores/settingsStore";
 import { importCalendarData } from "@/services/calendarService";
-import { ExportInstructions } from "./components/ExportInstructions";
 import { googleCalendarService } from "./services/googleCalendarService";
 import { importGoogleCalendar } from "./services/calendarService";
+import { Schedule } from "./components/Schedule";
+import { Insights } from "./components/Insights";
+import { cn } from "@/lib/utils";
 
 function App() {
-  const {
-    meetings: storedMeetings,
-    setMeetings,
-    clearAllData,
-    targetHours,
-  } = useSettingsStore();
-  const [meetings, setLocalMeetings] = useState<Meeting[]>(storedMeetings);
-  const [isLoading, setIsLoading] = useState(true);
-  const [stats, setStats] = useState<MeetingStats>({
-    totalHours: storedMeetings.reduce(
-      (acc, meeting) => acc + meeting.duration,
-      0
-    ),
-    targetHours,
-    availableHours: Math.max(
-      0,
-      targetHours -
-        storedMeetings.reduce((acc, meeting) => acc + meeting.duration, 0)
-    ),
-    overHours: Math.max(
-      0,
-      storedMeetings.reduce((acc, meeting) => acc + meeting.duration, 0) -
-        targetHours
-    ),
-  });
-  const [useMockData, setUseMockData] = useState<boolean>(false);
+  const { meetings, setMeetings, clearAllData } = useSettingsStore();
   const [isImporting, setIsImporting] = useState(false);
-
-  useEffect(() => {
-    updateStats(meetings);
-  }, [targetHours, meetings]);
-
-  const updateStats = (meetings: Meeting[]) => {
-    const { targetHours } = useSettingsStore.getState();
-    const totalHours = meetings.reduce(
-      (acc, meeting) => acc + meeting.duration,
-      0
-    );
-    const overHours = Math.max(0, totalHours - targetHours);
-    const availableHours = Math.max(0, targetHours - totalHours);
-
-    setStats({
-      totalHours,
-      targetHours,
-      availableHours,
-      overHours,
-    });
-  };
-
-  const updateMeetings = useCallback(
-    (newMeetings: Meeting[]) => {
-      setLocalMeetings(newMeetings);
-      setMeetings(newMeetings);
-    },
-    [setMeetings]
+  const [useMockData, setUseMockData] = useState(false);
+  const [activeTab, setActiveTab] = useState<"schedule" | "insights">(
+    "schedule"
   );
 
   useEffect(() => {
-    try {
-      if (useMockData) {
-        const rankedMeetings = mockMeetings.map((meeting, index) => ({
-          ...meeting,
-          rank: index + 1,
-        }));
-        updateMeetings(rankedMeetings);
-        updateStats(rankedMeetings);
-      }
-    } catch (error) {
-      console.error("Error loading mock data:", error);
-    } finally {
-      setIsLoading(false);
+    googleCalendarService.initializeGoogleApi(
+      import.meta.env.VITE_GOOGLE_CLIENT_ID
+    );
+  }, []);
+
+  useEffect(() => {
+    if (useMockData) {
+      const rankedMeetings = mockMeetings.map((meeting, index) => ({
+        ...meeting,
+        rank: index + 1,
+      }));
+      setMeetings(rankedMeetings);
     }
-  }, [useMockData, updateMeetings]);
+  }, [useMockData, setMeetings]);
+
+  const handleClearData = () => {
+    if (
+      window.confirm(
+        "Are you sure you want to clear all data? This cannot be undone."
+      )
+    ) {
+      clearAllData();
+      setMeetings([]);
+    }
+  };
 
   const handleFileUpload = async (
     event: React.ChangeEvent<HTMLInputElement>
@@ -106,82 +59,20 @@ function App() {
 
     try {
       const importedMeetings = await importCalendarData(file);
-      updateMeetings(importedMeetings);
-      updateStats(importedMeetings);
+      setMeetings(importedMeetings);
     } catch (error) {
       console.error("Error importing calendar:", error);
-      // Add user feedback here (e.g., toast notification)
     }
   };
 
-  const handleMeetingAction = (action: string, meetingId: string) => {
-    const store = useSettingsStore.getState();
-    console.log("Current store state:", store);
-    console.log("Action:", action, "MeetingId:", meetingId);
-
-    const currentStatus = store.meetingStatus?.[meetingId] || {
-      needsCancel: false,
-      needsShorten: false,
-      needsReschedule: false,
-      prepRequired: false,
-    };
-
-    console.log("Current status:", currentStatus);
-
-    // Update the status based on the action
-    const newStatus = {
-      ...currentStatus,
-      needsCancel:
-        action === "cancel"
-          ? !currentStatus.needsCancel
-          : currentStatus.needsCancel,
-      needsShorten:
-        action === "shorten"
-          ? !currentStatus.needsShorten
-          : currentStatus.needsShorten,
-      needsReschedule:
-        action === "reschedule"
-          ? !currentStatus.needsReschedule
-          : currentStatus.needsReschedule,
-      prepRequired:
-        action === "prep"
-          ? !currentStatus.prepRequired
-          : currentStatus.prepRequired,
-    };
-
-    console.log("New status:", newStatus);
-    store.setMeetingStatus?.(meetingId, newStatus);
+  const handleMockData = () => {
+    if (!useMockData) {
+      setMeetings(mockMeetings);
+    } else {
+      setMeetings([]);
+    }
+    setUseMockData(!useMockData);
   };
-
-  const handleDragEnd = (result: DropResult) => {
-    if (!result.destination) return;
-
-    const items = Array.from(meetings);
-    const [reorderedItem] = items.splice(result.source.index, 1);
-    items.splice(result.destination.index, 0, reorderedItem);
-
-    const updatedItems = items.map((item, index) => ({
-      ...item,
-      rank: index + 1,
-    }));
-
-    updateMeetings(updatedItems);
-    updateStats(updatedItems);
-  };
-
-  useEffect(() => {
-    googleCalendarService.initializeGoogleApi(
-      import.meta.env.VITE_GOOGLE_CLIENT_ID
-    );
-  }, []);
-
-  if (isLoading) {
-    return (
-      <div className="min-h-screen flex items-center justify-center">
-        Loading...
-      </div>
-    );
-  }
 
   return (
     <div className="min-h-screen bg-background text-foreground">
@@ -207,26 +98,10 @@ function App() {
               <EnvelopeClosedIcon className="w-4 h-4 mr-2" />
               Send Email
             </Button>
-            <Button
-              variant="destructive"
-              onClick={() => {
-                if (
-                  window.confirm(
-                    "Are you sure you want to clear all data? This cannot be undone."
-                  )
-                ) {
-                  clearAllData();
-                  setLocalMeetings([]);
-                  updateStats([]);
-                }
-              }}
-            >
+            <Button variant="destructive" onClick={handleClearData}>
               Clear All Data
             </Button>
-            <Button
-              variant="outline"
-              onClick={() => setUseMockData(!useMockData)}
-            >
+            <Button variant="outline" onClick={handleMockData}>
               {useMockData ? "Use Real Data" : "Use Mock Data"}
             </Button>
             <Button
@@ -237,8 +112,7 @@ function App() {
                   setIsImporting(true);
                   await googleCalendarService.authenticate();
                   const meetings = await importGoogleCalendar();
-                  updateMeetings(meetings);
-                  updateStats(meetings);
+                  setMeetings(meetings);
                 } catch (error) {
                   console.error("Error importing from Google Calendar:", error);
                 } finally {
@@ -252,53 +126,32 @@ function App() {
           </div>
         </div>
 
-        <ExportInstructions />
-
-        <StatsPanel stats={stats} />
-
-        <DragDropContext onDragEnd={handleDragEnd}>
-          <Droppable droppableId="meetings">
-            {(provided) => (
-              <div
-                {...provided.droppableProps}
-                ref={provided.innerRef}
-                className="space-y-1"
-              >
-                {meetings.map((meeting, index) => {
-                  const runningTotal = meetings
-                    .slice(0, index + 1)
-                    .reduce((acc, m) => acc + m.duration, 0);
-
-                  const currentTargetHours =
-                    useSettingsStore.getState().targetHours;
-
-                  return (
-                    <Draggable
-                      key={meeting.id}
-                      draggableId={meeting.id}
-                      index={index}
-                    >
-                      {(provided) => (
-                        <div
-                          ref={provided.innerRef}
-                          {...provided.draggableProps}
-                          {...provided.dragHandleProps}
-                        >
-                          <MeetingCard
-                            meeting={meeting}
-                            isOverTarget={runningTotal > currentTargetHours}
-                            onAction={handleMeetingAction}
-                          />
-                        </div>
-                      )}
-                    </Draggable>
-                  );
-                })}
-                {provided.placeholder}
-              </div>
+        <div className="flex space-x-1 mb-6">
+          <button
+            onClick={() => setActiveTab("schedule")}
+            className={cn(
+              "px-4 py-2 text-sm font-medium rounded-md",
+              activeTab === "schedule"
+                ? "bg-primary text-primary-foreground"
+                : "hover:bg-muted"
             )}
-          </Droppable>
-        </DragDropContext>
+          >
+            Schedule
+          </button>
+          <button
+            onClick={() => setActiveTab("insights")}
+            className={cn(
+              "px-4 py-2 text-sm font-medium rounded-md",
+              activeTab === "insights"
+                ? "bg-primary text-primary-foreground"
+                : "hover:bg-muted"
+            )}
+          >
+            Insights
+          </button>
+        </div>
+
+        {activeTab === "schedule" ? <Schedule /> : <Insights />}
       </div>
     </div>
   );
